Move USB device backends to separate files
