+++ Cargo.toml
@@ -33,6 +33,8 @@ object = { version = "0.38", default-features = false, features = [
 ] }
 indicatif = "0.18"
 serialport = "4.7"
-libloading = "0.9"
 chrono = "0.4"
 clap-verbosity-flag = "3"
+
+[target.'cfg(all(target_os = "windows", target_arch = "x86"))'.dependencies]
+libloading = "0.9"
